# [coreinternal/pdatautil] Reduce allocations in Map/ValueHash functions