Use a buffer pool instead of making allocations each time. The remaining allocations are caused by `sort.Strings` and `pdata.BytesSlice.AsRaw`.

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/pdatautil
BenchmarkMapHashFourItems
BenchmarkMapHashFourItems-10                  	 3136358	       383.7 ns/op	     288 B/op	      11 allocs/op
BenchmarkMapHashEightItems
BenchmarkMapHashEightItems-10                 	 1588498	       777.5 ns/op	     456 B/op	      19 allocs/op
BenchmarkMapHashWithEmbeddedSliceAndMap
BenchmarkMapHashWithEmbeddedSliceAndMap-10    	  933703	      1295 ns/op	     784 B/op	      32 allocs/op
PASS
```

After:
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/pdatautil
BenchmarkMapHashFourItems
BenchmarkMapHashFourItems-10                  	 5163958	       226.0 ns/op	      24 B/op	       1 allocs/op
BenchmarkMapHashEightItems
BenchmarkMapHashEightItems-10                 	 2389960	       502.4 ns/op	      32 B/op	       2 allocs/op
BenchmarkMapHashWithEmbeddedSliceAndMap
BenchmarkMapHashWithEmbeddedSliceAndMap-10    	 1497609	       799.8 ns/op	      56 B/op	       3 allocs/op
PASS
```