@MovieStoreGuy 2023-01-17
This looks fine, the only concern I had which I checked isn't the case is trying to borrow the underlying slice which can lead to some weirdness on `Reset`.